front: refactor export feature to BoardWrapper menu
