+++ front/public/locales/de/operational-studies.json
@@ -149,7 +149,6 @@
       "delete": "Löschen",
       "deleteSelection": "Ausgewählte Züge löschen",
       "duplicate": "Duplizieren",
-      "export": "Exportieren",
       "exportSelection": "Ausgewählte Züge exportieren",
       "filterLabel": "Name, Label",
       "importTimetableItem": "Züge importieren",
